requirements.txt
sqlalchemy
pydantic
passlib
python-jose
fastapi
uvicorn
python-dotenv
